Move error models to the errors package
