+++ pkg/errors/errors_test.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/contiamo/go-base/pkg/models"
 	"github.com/stretchr/testify/require"
 )
 
@@ -12,16 +11,16 @@ func TestValidationErrorsToFieldErrorResponse(t *testing.T) {
 	cases := []struct {
 		name     string
 		errs     ValidationErrors
-		expected models.FieldErrorResponse
+		expected FieldErrorResponse
 	}{
 		{
 			name:     "Returns empty response when errors are nil",
-			expected: models.FieldErrorResponse{Errors: []models.FieldError{}},
+			expected: FieldErrorResponse{Errors: []FieldError{}},
 		},
 		{
 			name:     "Returns empty response when errors are empty",
 			errs:     ValidationErrors{},
-			expected: models.FieldErrorResponse{Errors: []models.FieldError{}},
+			expected: FieldErrorResponse{Errors: []FieldError{}},
 		},
 		{
 			name: "Returns errors in the response when errors are not empty",
@@ -29,18 +28,18 @@ func TestValidationErrorsToFieldErrorResponse(t *testing.T) {
 				"field1": errors.New("bad field1"),
 				"field2": errors.New("bad field2"),
 			},
-			expected: models.FieldErrorResponse{
-				Errors: []models.FieldError{
-					models.FieldError{
-						GeneralError: models.GeneralError{
-							Type:    models.FieldErrorType,
+			expected: FieldErrorResponse{
+				Errors: []FieldError{
+					FieldError{
+						GeneralError: GeneralError{
+							Type:    FieldErrorType,
 							Message: "bad field1",
 						},
 						Key: "field1",
 					},
-					models.FieldError{
-						GeneralError: models.GeneralError{
-							Type:    models.FieldErrorType,
+					FieldError{
+						GeneralError: GeneralError{
+							Type:    FieldErrorType,
 							Message: "bad field2",
 						},
 						Key: "field2",
@@ -54,11 +53,11 @@ func TestValidationErrorsToFieldErrorResponse(t *testing.T) {
 				"field1": errors.New("bad field1"),
 				"field2": nil,
 			},
-			expected: models.FieldErrorResponse{
-				Errors: []models.FieldError{
-					models.FieldError{
-						GeneralError: models.GeneralError{
-							Type:    models.FieldErrorType,
+			expected: FieldErrorResponse{
+				Errors: []FieldError{
+					FieldError{
+						GeneralError: GeneralError{
+							Type:    FieldErrorType,
 							Message: "bad field1",
 						},
 						Key: "field1",
