+++ pkg/errors/field_error.go
@@ -1,4 +1,4 @@
-package models
+package errors
 
 // FieldError represents a validation error
 type FieldError struct {
