+++ pkg/errors/general_error.go
@@ -1,4 +1,4 @@
-package models
+package errors
 
 // GeneralError represents a system error exposed to the user
 type GeneralError struct {
