+++ pkg/errors/general_error_response.go
@@ -1,4 +1,4 @@
-package models
+package errors
 
 // GeneralErrorResponse - General error response that usually has a very generic message
 type GeneralErrorResponse struct {
